+++ src/momento/cache_client_async.py
@@ -8,10 +8,10 @@ from momento import logs
 from momento.auth import CredentialProvider
 from momento.config import Configuration
 from momento.errors import UnknownException
+from momento.internal._utilities._data_validation import validate_eager_connection_timeout
 from momento.requests import CollectionTtl, SortOrder
 from momento.utilities.shared_sync_asyncio import (
     DEFAULT_EAGER_CONNECTION_TIMEOUT_SECONDS,
-    validate_eager_connection_timeout,
 )
 
 try:
